Move storage to helpers
